DeleteDialog and limit contest

+++ src/app/modules/panel/components/author/index-author/index-author.component.ts
@@ -4,6 +4,7 @@ import { Guid } from 'guid-typescript';
 import { AuthorService } from '../service/author.service';
 import { Author } from '../model/author';
 import { PaginationInstance } from 'ngx-pagination/dist/pagination-instance';
+import {ConfirmationService, ConfirmEventType, MessageService} from 'primeng/api';
 const tableName = 'Author';
 @Component({
   selector: 'app-index-author',
@@ -24,17 +25,35 @@ export class IndexAuthorComponent implements OnInit {
     currentPage: this.page,
     totalItems: this.total
   };
-  constructor(private authorService: AuthorService, private router: Router) { }
+  constructor(private authorService: AuthorService, private router: Router,
+    private confirmationService: ConfirmationService, private messageService: MessageService) { }
 
   ngOnInit(): void {
     this.showData(1);
   }
-  deleteBook(id: Guid){ 
-    this.authorService.delete(id).subscribe(res => { 
-         this.authors = this.authors.filter(item => item.id !== id);
-         console.log('Author deleted successfully!');
-         this.router.navigateByUrl('panel/author');
-    })
+  onDelete(id: Guid, elementName: string){
+    this.confirmationService.confirm({
+      message: `Do you want to delete ${elementName}?`,
+      header: 'Delete Confirmation',
+      icon: 'pi pi-info-circle',
+      accept: () => {
+            this.authorService.delete(id).subscribe(res => { 
+              this.authors = this.authors.filter(item => item.id !== id);
+              this.messageService.add({ key: 'myKey1', severity:'info', summary:'Confirmed', detail:'Record deleted'});
+              this.router.navigateByUrl('panel/author');
+        })
+      },
+      reject: (type: any) => {
+          switch(type) {
+              case ConfirmEventType.REJECT:
+                  this.messageService.add({ key: 'myKey1', severity:'error', summary:'Rejected', detail:'You have rejected'});
+              break;
+              case ConfirmEventType.CANCEL:
+                  this.messageService.add({ key: 'myKey1', severity:'warn', summary:'Cancelled', detail:'You have cancelled'});
+              break;
+          }
+      }
+  });
   }
   
   showData(page: any): void {
