+++ src/app/modules/panel/components/book/index-book/index-book.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Guid } from 'guid-typescript';
 import { PaginationInstance } from 'ngx-pagination/dist/pagination-instance';
+import { ConfirmationService, ConfirmEventType, MessageService } from 'primeng/api';
 import { Book } from '../model/book';
 import { BookService } from '../service/book.service';
 
@@ -27,27 +28,43 @@ export class IndexBookComponent implements OnInit {
     totalItems: this.total
   };
   
-  constructor(private bookService: BookService, private router: Router) { }
+  constructor(private bookService: BookService, private router: Router,
+    private confirmationService: ConfirmationService, private messageService: MessageService) { }
 
   ngOnInit(): void {
     this.showData(1);
   }
   
-  deleteBook(id: Guid){
-    this.bookService.delete(id).subscribe(res => {
-         this.books = this.books.filter(item => item.id !== id);
-         console.log('Book deleted successfully!');
-         this.router.navigateByUrl('panel/book');
-    })
+  onDelete(id: Guid, elementName: string){
+    this.confirmationService.confirm({
+      message: `Do you want to delete ${elementName}?`,
+      header: 'Delete Confirmation',
+      icon: 'pi pi-info-circle',
+      accept: () => {
+            this.bookService.delete(id).subscribe(res => {
+              this.books = this.books.filter(item => item.id !== id);
+              this.messageService.add({ key: 'myKey1', severity:'info', summary:'Confirmed', detail:'Record deleted'});
+              this.router.navigateByUrl('panel/book');
+        })
+      },
+      reject: (type: any) => {
+          switch(type) {
+              case ConfirmEventType.REJECT:
+                  this.messageService.add({ key: 'myKey1', severity:'error', summary:'Rejected', detail:'You have rejected'});
+              break;
+              case ConfirmEventType.CANCEL:
+                  this.messageService.add({ key: 'myKey1', severity:'warn', summary:'Cancelled', detail:'You have cancelled'});
+              break;
+          }
+      }
+  });
   }
   
   showData(page: any): void {
     this.page = page;
   this.loading = true;
-  debugger
     this.bookService.getAllByRoute(page, this.config.itemsPerPage)
         .subscribe( res => {
-          debugger
           this.config.currentPage = page;
           this.books = res.results;
           this.total = res.rowCount;
